fuck

+++ build.sh
@@ -8,14 +8,16 @@ elif [ "$1" = "pull" ]; then
     clear
     git reset --hard
     git pull
-else
+elif [ "$1" = "test" ]; then
     rm -rf ../AirShitTest
     mkdir ../AirShitTest
     cp -r asset ../AirShitTest
-    # find . -maxdepth 1 -type f ! -name "tt.txt" -exec cp {} ../AirShitTest \;
     cp *.java ../AirShitTest
     javac -d . *.java
     javac -d ../AirShitTest ../AirShitTest/*.java
+    java AirShit.Main
+else
+    javac -d . *.java
     java AirShit.Main
 fi
